Changed input variables linsolver_type and linsolver_verbosity from double to int (as required by the function init)

+++ examples/exp_variogram.cpp
@@ -79,8 +79,8 @@ int main(int argc, char** argv)
     double minpermSI = Dune::unit::convert::from(minperm, Dune::prefix::milli*Dune::unit::darcy);
     double z_tolerance = param.getDefault("z_tolerance", 0.0);
     double residual_tolerance = param.getDefault("residual_tolerance", 1e-8);
-    double linsolver_verbosity = param.getDefault("linsolver_verbosity", 0);
-    double linsolver_type = param.getDefault("linsolver_type", 1);
+    int linsolver_verbosity = param.getDefault("linsolver_verbosity", 0);
+    int linsolver_type = param.getDefault("linsolver_type", 1);
 
     if (ilen <= 0) {
         std::cerr << "Error: ilen (" << ilen << ") must be greater than zero\n";
